refactor: split __init__.py and deduplicate structuredClone polyfill
Backend:
- Extract security utilities to security.py (validate_and_sanitize_filename,
  detect_file_type, validate_magic_bytes, constants)
- Extract WebSocket handlers to websocket.py (settings CRUD, photo
  upload/delete, deep_merge)
- Slim __init__.py to setup/config/frontend only (~170 lines, was 661)
- Re-export security functions from __init__.py for backward compat

Frontend:
- Remove inline structuredClone polyfill from dashview-panel.js
  (was duplicated with utils/polyfills.js)
- Import initPolyfills() from single source of truth
- Update stale comments in polyfills.js and utils/index.js
